requirements.txt
praw
python-dotenv
nltk
pandas
matplotlib
streamlit
wordcloud
